% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{routes}
\alias{routes}
\title{Get Routes}
\usage{
routes(x, overview = "simplified", nc = 1, nq = 100, server, profile = "car")
}
\arguments{
\item{x}{dataframe with x_src, y_src, x_dst, y_dst in epsg:4326}

\item{overview}{overview (FALSE, "simplified" or "full)}

\item{nc}{number of CPU cores}

\item{nq}{number of queries per chunk}

\item{server}{server address}

\item{profile}{profil}
}
\value{
smg
}
\description{
A short description...
}
\examples{
\dontrun{
library(osrm)
apt <- read.csv(system.file("csv/apotheke.csv", package = "osrm"))
x <- cbind(apt[c(1:100,1:100,1:100, 1:100), c(2:3)],
           apt[c(rep(1,100), rep(10,100), rep(20,100), rep(30, 100)), c(2:3)])
server = "http://xxxxxxx:5000/"
r <- routes(x, nc = 8, nq = 50, server = server, overview = "full")
library(mapsf)
op <- par(mfrow = c(2,2))
mf_map(r[1:100,], col = "red")
mf_map(r[101:200,], col = "blue")
mf_map(r[201:300,], col = "darkgreen")
mf_map(r[301:400,], col = "purple")
par(op)
}
}
